+++ .gitignore
@@ -1,5 +1,2 @@
-# Dependency directories
-node_modules/
-
-# Personal notes
-local-notes.md
+node_modules/*
+dist/*
