+++ package.json
@@ -7,16 +7,25 @@
     "type": "git",
     "url": "git@github.com:stracker-phil/codeceptjs-pixelmatchhelper.git"
   },
+  "scripts": {
+    "test": "jest"
+  },
   "dependencies": {
     "fs": "0.0.1-security",
     "path": "^0.12.7",
     "pixelmatch": "^5.2.1",
-    "pngjs": "^6.0.0"
+    "pngjs": "^6.0.0",
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.18.6",
+    "@jest/globals": "^29.3.1",
     "codeceptjs": "^3.1.1",
+    "jest": "^29.3.1",
     "mocha": "^8.4.0",
-    "mochawesome": "^6.2.2"
+    "mochawesome": "^6.2.2",
+    "ts-jest": "^29.0.3"
   },
   "keywords": [
     "codecept.io",
